split erratum out of models
