+++ collectors/errata/core.py
@@ -11,7 +11,8 @@ from requests_gssapi import HTTPSPNEGOAuth
 
 from collectors.framework.models import CollectorMetadata
 from osidb.core import set_user_acls
-from osidb.models import Erratum, Tracker
+from osidb.dmodels.erratum import Erratum
+from osidb.models import Tracker
 
 from ..utils import BACKOFF_KWARGS, fatal_code
 from .constants import (
