+++ collectors/errata/tests/test_core.py
@@ -2,7 +2,8 @@ import pytest
 from django.utils import timezone
 
 from collectors.framework.models import CollectorMetadata
-from osidb.models import Affect, Erratum, Tracker
+from osidb.dmodels.erratum import Erratum
+from osidb.models import Affect, Tracker
 from osidb.tests.factories import (
     AffectFactory,
     ErratumFactory,
