+++ osidb/dmodels/__init__.py
@@ -6,6 +6,9 @@ from .ps_product import PsProduct
 from .ps_update_stream import PsUpdateStream
 
 __all__ = (
+    # Erratum cannot be added here
+    # as it would make a cycle with Tracker import
+    #
     # package versions cannot be added here
     # as it would make a cycle with Flaw import
     "Profile",
